Add initial implementation of WiFi, ESP-NOW, web server, LED control, and voltmeter functionalities
