+++ app/esp_daemon/arduino/ESP_Daemon/data/index.html
@@ -1,16 +1,12 @@
 <!DOCTYPE html>
 <html>
   <head>
+    <meta charset="UTF-8">
     <title>DIT Robotics</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="icon" type="image/png" href="favicon.png">
     <link rel="stylesheet" type="text/css" href="style.css">
-    <script src="https://code.highcharts.com/highcharts.js"></script>
-    <script>
-        if (typeof Highcharts === 'undefined') {
-            document.write('<script src="highcharts.js"><\/script>');
-        }
-    </script>
+    <script src="highcharts.js"></script>
   </head>
   <body>
     <div class="topnav">
