+++ app/esp_daemon/arduino/ESP_Daemon/lib/sensors/voltmeter.h
@@ -1,0 +1,14 @@
+#ifndef VOLTMETER_H
+#define VOLTMETER_H
+
+#include <Arduino.h>
+
+extern float Vbattf;
+
+void initVoltmeter();
+void voltmeter();
+void IRAM_ATTR onTimer();
+void voltmeterTask(void* pvParameters);
+String getSensorReadings();
+
+#endif
