+++ app/esp_daemon/arduino/ESP_Daemon/lib/web_server/web_server.h
@@ -1,0 +1,12 @@
+#ifndef WEB_SERVER_H
+#define WEB_SERVER_H
+
+#include <ESPAsyncWebServer.h>
+
+extern AsyncWebServer server;
+extern AsyncEventSource events;
+
+void initSPIFFS();
+void initWebServer();
+
+#endif
