+++ app/esp_daemon/arduino/ESP_Daemon/lib/wifi_config/wifi_config.h
@@ -1,0 +1,6 @@
+#ifndef WIFI_CONFIG_H
+#define WIFI_CONFIG_H
+
+void initWiFi();
+
+#endif
