+++ app/esp_daemon/arduino/ESP_Daemon/platformio.ini
@@ -20,6 +20,7 @@ lib_deps =
 	me-no-dev/ESPAsyncWebServer@^3.0.6
 	tzapu/WiFiManager@^2.0.17
 	https://github.com/micro-ROS/micro_ros_platformio
+build_flags = -Iinclude
 upload_port = /dev/ttyACM0
 monitor_port = /dev/ttyACM0
 monitor_speed = 115200
